Renamed folder User -> UserTests in test solution
